Extract SAPDownloadedPackage helpers
