+++ src/xcode/ENA/ENATests/Helper/MockKeyStore.swift
@@ -1,0 +1,34 @@
+//
+// Corona-Warn-App
+//
+// SAP SE and all other contributors
+// copyright owners license this file to you under the Apache
+// License, Version 2.0 (the "License"); you may not use this
+// file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+//
+
+@testable import ENA
+import Foundation
+import CryptoKit
+
+struct MockKeyStore: PublicKeyStore {
+	let keys: [String: P256.Signing.PublicKey]
+
+	func publicKey(for bundleID: String) throws -> P256.Signing.PublicKey {
+		guard let key = keys[bundleID] else {
+			throw KeyError.environmentError
+		}
+
+		return key
+	}
+}
